Fix a quite obscure bug causing playback of modified 24-bit files to fail.
Turns out the bug was in chunk_read_array. If the requested size was not
an even multiple of the sample size, the datasource_read_array function
would return less than the requested size, which is interpreted as an error
and chunk_read_array returned 0. If this was on the last datasource,
however, this would be interpreted as an EOF and the partial size would be
returned, this was why it worked for unmodified files.