Translate each pagagraph of help text separately
With the earlier scheme, each help page was generated as one big string,
which is translated at runtime. This means that any small change in the
help text will lead to the whole page getting out of date.

Change gendocs.sh so that we generate an array of lines for each help page.
In the help display code, these lines are translated individually and then
concatenated into the page contents.