mainwindow.c: Make view move keys also move playback if follow-mode enabled
When we have follow mode enabled and playback is running, the view changing
keys left/right-arrow and home/end did not have any effect since the view
jumps back shortly after the key has taken effect.

Now we instead move the playback when this happens. To reflect this change,
the follow mode tooltip is changed to "Keep view and playback together"