document.c: Fix invalid state when playing very small files
When calling player_play, for some drivers it is possible for the playback
to finish during the call and the notify callback will then be called
immediately. This happens especially for Jack since the playback is done
in another thread. Because playing_document is set after the call to
player_play, this caused the program to get into an invalid state, where
the player is stopped but playing_document is != NULL.