Add unbiased PCM-to-float conversion routines and use by default
The values used in standard PCM wav files are slightly assymmetrical, for
example 16-bit samples range between -32768 and +32767. The existing
routines converted these to normalized floating-point by adding 0.5 and
then dividing by 32767.5. The reverse was done to convert back.

This has the advantages of preserving sample values across conversions and
avoiding clipping. It also gives consistency between signed and unsigned
types (converting signed->unsigned->float gives the same result).

A downside with this conversion method however is that the PCM value 0
does not correspond to the exact floating-point value 0.0. It also doesn't
match how other programs and libraries, for example Jack, SoX and
libsndfile, convert between PCM and float.

Therefore new routines that just scale are added and used by default
instead. In the 16-bit case above the routine just divides by 32767. When
converting back, the floating-point value is multiplied by the same value.

This makes the PCM and floating-point zero levels match, but the minimum
value and lower clipping point will no longer be exactly -1.0 (-1.00003
for 16-bit). This should normally not be a problem with any of the
built-in effects or Ladspa plugins, if it is you will have to normalize
the file before processing it.

The lower level is different for different bit-widths (for 24-bit signed
it is -1.0000001), therefore a scaling factor is added when converting to
a higher bit width to avoid introducing clipping in the process.

The new behavior is default, but the old one can be enabled through the
Settings dialog, by setting Normalization mode to "Biased".