dither_convert: Use scale factor table to avoid numerical problems
In case data is converted to float and back again without change, we want
dithering to have no effect. To avoid rounding problems we need to have
the maximum dither amount a tiny bit less than half an LSB. A separate
test program was used to search for the optimum constants. In case of
conversion from float to 24/32-bit PCM, we disable dithering.