Fix dithering noise
Triangular dithering with +/- 0.5 LSB range is not statistically correct to
use as dither noise, It should be either +/- 0.5 LSB rectangular or
+/- 1 LSB triangular to get the right expected value.

We replace the dithering with a +/- 0.5 LSB rectangular dither. The reason
for choosing this dither is that it is the variant which adds the least
amount of noise and it also preserves bit-exactness in case the values are located
exactly on the quantization levels. The latter can happen for example when
loading and saving lower-bit sources and some samples are unchanged by an
effect.